+++ changelog.rst
@@ -3,10 +3,13 @@ Changelog
 
 Last change: 22-FEB-2024 MTS
 
-0.6.9
------
+0.6.9 - 0.6.10
+--------------
 - Added the option to draw polygon picks in Picasso: Render
 - Save pick properties in Picasso: Render saves areas of picked regions in nm^2
+- Calibration .yaml file saves number of frames and step size in nm
+- `picasso.lib.merge_locs` function can merge localizations from multiple files
+
 
 0.6.6 - 0.6.8
 -------------